ui: added mintnft button, fixed typography, recfator code
